ch4/rma: store is_local info in callback-created request
Some RMA callback internally creates request and associates it with
ACK reply. ACK reply is performed in the completion callback function
pointer, and checks the locality of the peer. The current code
incorrectly assumed the is_local info is already cached in the request
and used it for locality check. This patch adds the caching step in
callback, thus fixes the issue.